More steps in cleaning
